+++ LittleEngine/Source/Engine/Physics/CollisionManager.cpp
@@ -56,9 +56,9 @@ namespace LittleEngine {
 	}
 
 	void CollisionManager::Cleanup() {
-		int count = colliders.size();
+		int count = static_cast<int>(colliders.size());
 		Utils::EraseNullWeakPtrs<Collider>(colliders);
-		int diff = count - colliders.size();
+		int diff = count - static_cast<int>(colliders.size());
 		if (diff > 0) {
 			Logger::Log(*this, "Removed " + std::to_string(diff) + " stale Colliders ", Logger::Severity::Debug);
 		}
